[feature/AB#49323-new-renewal-cert] AB#52263 Created unit tests to verify that the Intoduction and Signature text corresponds with the UX.
