+++ CDTS Blazor/Resources/EmailTemplates/Parameters/SubmissionEmailTemplateSignatureEnglish.html
@@ -2,8 +2,8 @@
 <p>Our team will review your documents within the next 10 working days. We will ensure they meet the requirements for your selected certificate.</p>
 <p>If you meet the requirements, you can expect to receive your certificate:</p>
 <ul>
-    <li>Within 10 working days for an Examiner's Certificate</li>
-    <li>Within 120 working days for a Minister's Certificate</li>
+<li>Within 10 working days for an Examiner's Certificate</li>
+<li>Within 120 working days for a Minister's Certificate</li>
 </ul>
 <p>We will contact you by the phone number or email address your provided if we need any more information.</p>
 <p>If you have questions or concerns, contact your local <a href="https://tc.canada.ca/en/corporate-services/regions" target="_blank">Transport Canada Service Centre</a>. Be sure to have your confirmation number ready to share.</p>
