+++ CDTS Blazor/Resources/EmailTemplates/Parameters/SubmissionEmailTemplateSignatureFrench.html
@@ -2,8 +2,8 @@
 <p>Notre équipe examinera vos documents dans un délai de 10 jours ouvrables. Nous nous assurerons qu’ils répondent aux exigences du brevet que vous avez sélectionné.</p>
 <p>Si vous remplissez les conditions requises, vous pouvez vous attendre à recevoir votre brevet:</p>
 <ul>
-    <li>Dans un délai de 10 jours ouvrables pour brevet de l’examinateur</li>
-    <li>Dans un délai de 120 jours ouvrables pour un brevet du ministre</li>
+<li>Dans un délai de 10 jours ouvrables pour brevet de l’examinateur</li>
+<li>Dans un délai de 120 jours ouvrables pour un brevet du ministre</li>
 </ul>
 <p>Nous communiquerons avec vous au moyen du numéro de téléphone ou de l’adresse électronique que vous nous avez fournis, si nous avons besoin de plus de renseignements.</p>
 <p>Si vous avez des questions ou des préoccupations, communiquez avec le <a href="https://tc.canada.ca/fr/services-generaux/regions" target="_blank">Centre de services de Transports Canada</a> de votre région. Assurez-vous d’avoir votre numéro de confirmation à portée de main.</p>
