+++ CDTS Blazor/appsettings.json
@@ -39,7 +39,7 @@
     "EmailNotificationPath": "api/v1/notifications?overrideEmailRecipientsSafeguard=true",
     "EmailNotificationTemplatePath": "api/v1/notification-template",
     "FileAttachmentsPath":  "api/v1/file-attachments",
-    "EmailSubmissionTemplateName": "Seafarers_Document_Submission_Email_TESTING",
+    "EmailSubmissionTemplateName": "Seafarers_Document_Submission_Email",
     "SeafarerCertificationServiceName": "CSF",
     "ReplyEmail": "donotreply-nepasrepondre@tc.gc.ca",
     "ServiceRequestId": "13844",
